Graphic project for Breathing Exercise
This is a short graphical simulator program for performing the breathing exercises that we so often prescribe to our patients in real time using color therapy. The user only needs to observe carefully and perform simple actions of inhaling and exhaling for a few minutes. This will sequentially print by the program and run a full cycle of deep breathing exercises, then change the background, with thanksgiving you for good job gradually end the workout after a long exhale with wishes for good health.